+++ src/vcl-comps/devtools/Workspace.js
@@ -371,6 +371,7 @@ $(["ui/Form"], {
             $("vcl/ui/Tab", { text: locale("Navigator"), control: "navigator", selected: true }),
             $("vcl/ui/Tab", { text: locale("Open Tabs"), control: "openTabs" }),
             $("vcl/ui/Tab", { text: locale("Console"), control: "console" }),
+            $("vcl/ui/Tab", { text: locale("Scratch"), control: "scratch" }),
             $("vcl/ui/Tab", { text: locale("Outline"), control: "outline" }),
             $("vcl/ui/Tab", { text: locale("Bookmarks"), control: "bookmarks", visible: false }),
             $("vcl/ui/Tab", { text: locale("Search"), control: "search-panel", visible: false }),
@@ -382,6 +383,8 @@ $(["ui/Form"], {
         $(["./Outline"], "outline", { _align: "client", visible: false }),
         $(["./OpenTabs"], "openTabs", { visible: false }),
         $(["./Console"], "console", { visible: false }),
+        
+        $("vcl/ui/Ace", "scratch", { align: "client", visible: false }),
 
         $("vcl/ui/Panel", "search-panel", { align: "client", visible: false }),
         $("vcl/ui/Panel", "inspector-panel", { align: "client", visible: false })