+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 2020-01-24 / 1.0.111
+- Double click `devtools/Editor<folder>#editors-tabs` to create new resource
+
+### 2020-01-22 / 1.0.110
++ cavalion-js@1.0.35
++ cavalion-vcl@1.0.70
+
+### 2020-01-10 / 1.0.109
++ cavalion-code-blocks@1.0.12
++ cavalion-vcl@1.0.69
+
 ### 2020-01-10 / 1.0.108
 	+ cavalion-code-blocks@1.0.10
 - Finetuning and bugfixing for a demo/collab with TO